+++ src/repositories/__init__.py
@@ -1,0 +1,1 @@
+"""Repositories do FlowLog: encapsulam o SQL por entidade."""